requirements.txt
Flask==2.3.3
python-dateutil==2.8.2
Werkzeug==2.3.7
